move PageRankOutput to computer-algorithm
